Changed to training off the observed state that's output from the Env instead of the screenshots.
Very much in progress. This works much better and is a fine example of how to train something. I'd like it if the screenshots worked but it seems like something was wrong.

Another major change was to make the reward for failure -1. It was previously always 1 which was confusing to the model.